Extract module `rustc_session::config::print_request`
